Separate Beebjit call
